refactor: reduce branch lint noise
